# packages/api-client/openapi/reboxe.yaml
openapi: 3.1.0
info:
  title: REBOXE API
  version: 1.4.0
servers:
  - url: http://127.0.0.1:3001
paths:
...
  /v1/shops/{shopId}/listings:
    parameters:
      - $ref: "#/components/parameters/ShopId"
    get:
      operationId: listShopListings
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Shop listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Listing" }
    post:
      operationId: createListingDraft
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListing"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /v1/shops/{shopId}/return-imports/preview:
    post:
      operationId: previewReturnManifest
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Parsed and allowlisted return manifest preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnManifestPreview"
  /v1/shops/{shopId}/return-imports/{batchId}/commit:
    post:
      operationId: commitReturnManifest
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/BatchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [idempotencyKey]
              properties:
                idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
      responses:
        "201":
          description: Idempotently committed return packages and source lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitReturnManifestResult"
  /v1/shops/{shopId}/listings/{listingId}/publish:
    post:
      operationId: publishListing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/ListingId"
      responses:
        "201":
          description: Policy evaluated; listing activated or sent to manual review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishListingResult"
  /v1/shops/{shopId}/listings/{listingId}/images/init:
    post:
      operationId: createCatalogImageUploadIntent
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatalogImageUpload"
      responses:
        "201":
          description: Signed catalog image upload intent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogImageUploadIntent"
  /v1/shops/{shopId}/listings/{listingId}/images/complete:
    post:
      operationId: completeCatalogImageUpload
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [key]
              properties:
                key: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "201":
          description: Uploaded image metadata verified and attached to the draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /v1/shops/{shopId}/listings/{listingId}:
    patch:
      operationId: updateListingDraft
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/ListingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingDraft"
      responses:
        "200":
          description: Draft updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /v1/listings/{listingId}:
    get:
      operationId: getPublicListing
      parameters:
        - $ref: "#/components/parameters/ListingId"
      responses:
        "200":
          description: Public listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicListing"
  /v1/shops/{shopId}:
    get:
      operationId: getPublicShop
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Public shop profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicShop"
  /v1/shops/{shopId}/reviews:
    get:
      operationId: listShopReviews
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Public shop reviews, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShopReview" }
  /v1/shops/{shopId}/reviews/mine:
    get:
      operationId: getMyShopReview
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Current actor's review, if any
          content:
            application/json:
              schema: { anyOf: [{ $ref: "#/components/schemas/ShopReview" }, { type: "null" }] }
    put:
      operationId: upsertShopReview
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertShopReview" }
      responses:
        "200":
          description: Review created or updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopReview" }
  /v1/shops/{shopId}/reviews/eligibility:
    get:
      operationId: getShopReviewEligibility
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "200":
          description: Whether the current actor may review the shop
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopReviewEligibility" }
        "401": { description: Authentication required }
        "404": { description: Active shop not found }
  /v1/listings:
    get:
      operationId: listPublicListings
      parameters:
        - name: cursor
          in: query
          schema: { type: string, maxLength: 512 }
        - name: q
          in: query
          schema: { type: string, maxLength: 180 }
        - name: category
          in: query
          schema: { type: string, maxLength: 80 }
        - name: shopId
          in: query
          schema: { type: string, maxLength: 80 }
        - name: sort
          in: query
          schema: { type: string, enum: [newest, price_asc, price_desc], default: newest }
      responses:
        "200":
          description: Public listings page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicListingPage"
  /v1/shops/{shopId}/return-packages/scan:
    post:
      operationId: scanReturnPackage
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: shopId, required: true, schema: { type: string } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ScanReturnPackage" } } }
      responses: { "201": { description: Idempotent package-backed draft } }
  /v1/shops/{shopId}/return-packages/listings/batch:
    post:
      operationId: batchCreatePackageListings
      security: [{ bearerAuth: [] }]
      parameters: [{ in: path, name: shopId, required: true, schema: { type: string } }]
      responses: { "201": { description: Per-package results } }
  /v1/admin/listings/reviews:
    get:
      operationId: listPendingListingReviews
      security: [{ bearerAuth: [] }]
      responses: { "200": { description: AAL2 moderation queue } }
  /v1/admin/listings/{listingId}/decision:
    post:
      operationId: decideListingReview
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: listingId, required: true, schema: { type: string } }
        - { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
      responses: { "200": { description: Idempotent moderation decision } }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ShopId:
      name: shopId
      in: path
      required: true
      schema: { type: string }
    ListingId:
      name: listingId
      in: path
      required: true
      schema: { type: string }
    BatchId:
      name: batchId
      in: path
      required: true
      schema: { type: string }
  schemas:
...
    CreateListing:
      type: object
      required: [title, categoryId, conditionGrade, conditionNotes, price, weightGram]
      properties:
        title: { type: string, minLength: 3, maxLength: 180 }
        description: { type: string, maxLength: 5000 }
        categoryId: { type: string }
        conditionGrade: { type: string, enum: [NEW_SEALED, LIKE_NEW_99, GOOD, FAIR, DEFECT] }
        conditionNotes: { type: string, minLength: 3, maxLength: 2000 }
        price: { type: integer, minimum: 1 }
        weightGram: { type: integer, minimum: 1, maximum: 100000 }
    UpdateListingDraft:
      $ref: "#/components/schemas/CreateListing"
      unevaluatedProperties: false
    CreateCatalogImageUpload:
      type: object
      additionalProperties: false
      required: [mimeType, sizeBytes]
      properties:
        mimeType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        sizeBytes: { type: integer, minimum: 1, maximum: 5242880 }
...
    CatalogImageUploadIntent:
      type: object
      required: [key, uploadUrl, expiresAt, headers]
      properties:
        key: { type: string }
        uploadUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        headers:
          type: object
          additionalProperties: { type: string }
...
    ListingImage:
      type: object
      additionalProperties: false
      required: [key, url, width, height]
      properties:
        key: { type: string }
        url: { type: string, format: uri }
        width: { type: integer, minimum: 1 }
        height: { type: integer, minimum: 1 }
    Listing:
      allOf:
        - $ref: "#/components/schemas/CreateListing"
        - type: object
          required: [id, shopId, shopDisplayName, images, status, publishedAt, createdAt]
          properties:
            id: { type: string }
            shopId: { type: string }
            shopDisplayName: { type: string }
            images:
              type: array
              maxItems: 6
              items: { $ref: "#/components/schemas/ListingImage" }
            status: { type: string, enum: [DRAFT, PENDING_REVIEW, ACTIVE, HIDDEN_BY_FUND, RESERVED, SOLD, RELISTABLE, SUSPENDED, DELISTED] }
            publishedAt: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
            createdAt: { type: string, format: date-time }
    ListingPolicyResult:
      type: object
      additionalProperties: false
      required: [outcome, policyLevel, policyVersion, message]
      properties:
        outcome: { type: string, enum: [ACTIVE, PENDING_REVIEW] }
        policyLevel: { anyOf: [{ type: string, enum: [BANNED, MANUAL_REVIEW, DISCLOSURE] }, { type: "null" }] }
        policyVersion: { anyOf: [{ type: string }, { type: "null" }] }
        message: { type: string }
    PublishListingResult:
      type: object
      additionalProperties: false
      required: [listing, policy]
      properties:
        listing: { $ref: "#/components/schemas/Listing" }
        policy: { $ref: "#/components/schemas/ListingPolicyResult" }
    ReturnManifestLineDraft:
      type: object
      additionalProperties: false
      required: [sourceItemRef, sourceQuantity, productName, productImageUrls, reboxeCategoryId]
      properties:
        sourceItemRef: { type: string }
        sourceSku: { type: string }
        sourceQuantity: { type: integer, minimum: 1 }
        productName: { type: string }
        variantName: { type: string }
        brand: { type: string }
        sourceCategory: { type: string }
        originalUnitPriceVnd: { type: integer, minimum: 1 }
        returnReason: { type: string }
        productImageUrls:
          type: array
          items: { type: string, format: uri }
        reboxeCategoryId: { type: string }
    ReturnManifestDraft:
      type: object
      additionalProperties: false
      required: [source, sourcePlatform, sourceTrackingNo, packageListingPriceVnd, lines]
      properties:
        source: { type: string, enum: [SPREADSHEET, PLATFORM_API] }
        sourcePlatform: { type: string, enum: [SHOPEE, TIKTOK] }
        sourceTrackingNo: { type: string }
        sourceOrderRef: { type: string }
        sourceReturnRef: { type: string }
        returnedAt: { type: string, format: date-time }
        packageWeightGram: { type: integer, minimum: 1, maximum: 100000 }
        packageDimensionsCm:
          type: object
          additionalProperties: false
          required: [length, width, height]
          properties:
            length: { type: number, exclusiveMinimum: 0 }
            width: { type: number, exclusiveMinimum: 0 }
            height: { type: number, exclusiveMinimum: 0 }
        packageListingPriceVnd: { type: integer, minimum: 1 }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ReturnManifestLineDraft" }
    ReturnManifestIssueCode:
      type: string
      enum: [INVALID_FIELD, INVALID_CATEGORY, PACKAGE_FIELD_CONFLICT, DUPLICATE_SOURCE_ITEM_REF]
    ReturnManifestPreviewRow:
      type: object
      additionalProperties: false
      required: [rowIndex, packageGroup, warningCodes, errorCodes]
      properties:
        rowIndex: { type: integer, minimum: 1 }
        packageGroup: { type: string }
        warningCodes:
          type: array
          items: { $ref: "#/components/schemas/ReturnManifestIssueCode" }
        errorCodes:
          type: array
          items: { $ref: "#/components/schemas/ReturnManifestIssueCode" }
    ReturnManifestPreview:
      type: object
      additionalProperties: false
      required: [batchId, rows, drafts, canCommit]
      properties:
        batchId: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/ReturnManifestPreviewRow" }
        drafts:
          type: array
          items: { $ref: "#/components/schemas/ReturnManifestDraft" }
        canCommit: { type: boolean }
    CommitReturnManifestResult:
      type: object
      additionalProperties: false
      required: [batchId, packageIds, lineCount]
      properties:
        batchId: { type: string }
        packageIds:
          type: array
          items: { type: string }
        lineCount: { type: integer, minimum: 0 }
    PublicListing:
      type: object
      required: [id, shopId, shopDisplayName, title, categoryId, conditionGrade, conditionNotes, price, images, publishedAt, createdAt]
      properties:
        id: { type: string }
        shopId: { type: string }
        shopDisplayName: { type: string }
        title: { type: string, minLength: 3, maxLength: 180 }
        description: { type: string, maxLength: 5000 }
        categoryId: { type: string }
        conditionGrade: { type: string, enum: [NEW_SEALED, LIKE_NEW_99, GOOD, FAIR, DEFECT] }
        conditionNotes: { type: string, minLength: 3, maxLength: 2000 }
        price: { type: integer, minimum: 1 }
        images:
          type: array
          maxItems: 6
          items: { $ref: "#/components/schemas/ListingImage" }
        publishedAt: { anyOf: [{ type: string, format: date-time }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
        availableQuantity: { type: integer, enum: [0, 1] }
        package:
          type: object
          properties:
            disclosure: { type: string, enum: [UNOPENED_UNINSPECTED] }
            sealStatus: { type: string, enum: [INTACT, DAMAGED, UNKNOWN] }
            manifestSummary:
              type: object
              properties:
                lineCount: { type: integer, minimum: 1 }
                unitCount: { type: integer, minimum: 1 }
            lines:
              type: array
              items:
                type: object
                properties:
                  productName: { type: string }
                  variantName: { anyOf: [{ type: string }, { type: "null" }] }
                  brand: { anyOf: [{ type: string }, { type: "null" }] }
                  quantity: { type: integer, minimum: 1 }
                  categoryId: { type: string }
                  originalUnitPriceVnd: { anyOf: [{ type: integer, minimum: 1 }, { type: "null" }] }
    ScanReturnPackage:
      type: object
      additionalProperties: false
      required: [scannedCode, codeType, platformHint]
      properties:
        scannedCode: { type: string, minLength: 1, maxLength: 200 }
        codeType: { type: string, enum: [ORDER_SN, TRACKING_NO, UNKNOWN] }
        platformHint: { anyOf: [{ type: string, enum: [SHOPEE, TIKTOK] }, { type: "null" }] }
...
    PublicListingPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/PublicListing" }
        nextCursor: { anyOf: [{ type: string }, { type: "null" }] }
    PublicShop:
      type: object
      additionalProperties: false
      required: [id, displayName, description, avatarUrl, verified, activeListingCount, averageRating, reviewCount, location, createdAt]
      properties:
        id: { type: string }
        displayName: { type: string }
        description: { anyOf: [{ type: string }, { type: "null" }] }
        avatarUrl: { anyOf: [{ type: string, format: uri }, { type: "null" }] }
        verified: { type: boolean }
        activeListingCount: { type: integer, minimum: 0 }
        averageRating: { anyOf: [{ type: number, minimum: 1, maximum: 5 }, { type: "null" }] }
        reviewCount: { type: integer, minimum: 0 }
        location: { anyOf: [{ type: string }, { type: "null" }] }
        createdAt: { type: string, format: date-time }
    UpsertShopReview:
      type: object
      additionalProperties: false
      required: [rating, content]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        content: { type: string, minLength: 3, maxLength: 1000 }
    ShopReview:
      allOf:
        - $ref: "#/components/schemas/UpsertShopReview"
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    ShopReviewEligibility:
      type: object
      additionalProperties: false
      required: [eligible, reason]
      properties:
        eligible: { type: boolean }
        reason:
          anyOf:
            - { type: string, enum: [COMPLETED_ORDER_REQUIRED, SHOP_MEMBER] }
            - { type: "null" }